requirements.txt
Django==3.2.5
djangorestframework==3.12.4
flake8==6.1.0
gunicorn==21.2.0
mccabe==0.7.0
mysql-connector-python==8.0.32
packaging==23.1
protobuf==3.20.3
psycopg2==2.9.7
pyasn1==0.5.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3
rsa==4.9
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
